+++ data/store/src/lib.rs
@@ -1,5 +1,6 @@
 #[macro_use]
 extern crate bson;
+extern crate chrono;
 #[macro_use]
 extern crate error_chain;
 #[macro_use]