+++ replicante/src/components/mod.rs
@@ -12,10 +12,14 @@ use super::Result;
 use super::metrics::COMPONENTS_ENABLED;
 
 
-pub mod discovery;
-pub mod webui;
+mod discovery;
+mod grafana;
+mod webui;
+
+pub use self::discovery::Config as DiscoveryConfig;
 
 use self::discovery::DiscoveryComponent as Discovery;
+use self::grafana::Grafana;
 use self::webui::WebUI;
 
 
@@ -39,7 +43,7 @@ macro_rules! component_wait {
 
 /// Helper function to keep `Components::new` simpler in the presence of optional components.
 fn component_new<C, F>(
-    component: &str, mode: &str, enabled: bool, logger: Logger, factory: F
+    component: &str, mode: &str, enabled: bool, logger: &Logger, factory: F
 ) -> Option<C>
     where F: FnOnce() -> C,
 {
@@ -68,6 +72,7 @@ fn component_new<C, F>(
 /// [`JoinHandle`]: std/thread/struct.JoinHandle.html
 pub struct Components {
     discovery: Option<Discovery>,
+    grafana: Option<Grafana>,
     webui: Option<WebUI>,
 }
 
@@ -75,19 +80,23 @@ impl Components {
     /// Creates and configures components.
     pub fn new(config: &Config, logger: Logger, interfaces: &mut Interfaces) -> Result<Components> {
         let discovery = component_new(
-            "discovery", "required", config.components.discovery(), logger.clone(), || {
+            "discovery", "required", config.components.discovery(), &logger, || {
                 Discovery::new(
                     config.discovery.clone(), config.events.snapshots.clone(),
                     Duration::from_secs(config.timeouts.agents_api), logger.clone(), interfaces
                 )
             }
         );
+        let grafana = component_new(
+            "grafana", "optional", config.components.grafana(), &logger,
+            || Grafana::new(interfaces)
+        );
         let webui = component_new(
-            "webui", "optional", config.components.webui(), logger.clone(),
-            || WebUI::new(interfaces)
+            "webui", "optional", config.components.webui(), &logger, || WebUI::new(interfaces)
         );
         Ok(Components {
             discovery,
+            grafana,
             webui,
         })
     }
@@ -106,6 +115,7 @@ impl Components {
     /// Performs any final configuration and starts background threads.
     pub fn run(&mut self) -> Result<()> {
         component_run!(self.discovery.as_mut());
+        component_run!(self.grafana.as_mut());
         component_run!(self.webui.as_mut());
         Ok(())
     }
@@ -113,6 +123,7 @@ impl Components {
     /// Waits for all interfaces to terminate.
     pub fn wait_all(&mut self) -> Result<()> {
         component_wait!(self.discovery.as_mut());
+        component_wait!(self.grafana.as_mut());
         component_wait!(self.webui.as_mut());
         Ok(())
     }